apps/_examples/demo-backend/app/pyproject.toml
[project]
name = "demo-backend"
version = "1.0.0"
description = "FastAPI demo backend for GitOps deployment"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.6",
    "psycopg2-binary>=2.9.9",
    "prometheus-client>=0.21.0",
    "pydantic>=2.9.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
    "ruff>=0.5.0",
    "mypy>=1.11.0",
]
